Don't set sensitivity filenames for each frame
We need to calculate the sensitivity for each frame (and each subset, if desired). The sensitivities for all the frames are then joined together via a weighted mean and the result is saved. After this, the sensitivity for each of the frames is not required, so no need to save it. Just save the sensitivity of the parametric objective function.